Add detection MWE as base for barebones SageMic
This is the exact same code that is in the branch for the MWE of
detection from a live microphone audio stream.  It seems like a
natural starting point for building out the barebones implementation
of the SageMic code.  Might want to rename this module later
(particularly if it gets designed as a client to a centralized server
program running on a SageBRUSH server) but for the time being this
will work.