add options to change input device
I was having issues with my default mic on my laptop, so I am
unable to use it to actually do inference from there. Because the
script defaults to the default sound device, I couldn't get it to
actually infer anything. I needed to change the input device, and
so I attached an audiomoth mic to my computer. This enabled it to
be on the sound device list, but I needed to add options for the
user to select a different sound device. Rather than hardcode this
in the feature test, I felt it made sense to give the user the option
to set it since the device could change for each users computer. One
also can't hardcode the name of the audiomoth device as the device
necessarily because the name of the usb mic changes depending on
the sample rate you set on the audiomoth mic firmware. IE, the name
of my audiomoth device is MCS 48kHz AudioMoth USB Microphone, which
would change if I changed the sample rate.