allow streaming from ip for audio rather than local sounddevices
the sagemic_local can use the local sounddevices to stream audio
and run inference, but we need to be able to stream audio to run
inference on a machine at the lab. the
sagemic_stream script is run on the machine doing inference, and
is accessing the audio stream from a pi with a mic that's hosting
a icecast server with a darkice audio stream. its an http
stream accessible on the browser