add ability to store the wav file of interest
not fully clean, but basic functionality added. will store the
wav file associated with a detection with the time and name
of detected bird and confidence. need to figure out a good
filesystem and debug audio weirdness. could be anything from
mismatched sample rate on the stream side to having a crappy mic,
but its essentially working