add better storage for detections
copying how birdnetpi stores detections, now each time its run it
will check for an existing folder in your basepath that marks the
current date, and if it doesnt exist it will create it. all
detection clips for the day will be stored in that days folder with
the filenames containing the time, the name of the bird, and
the confidence value of the detection. I think this should be
modified to check on startup and only once- but this would mean
the program would need to reboot once per day around midnight.